test(libs/tests/unit/backup-old-path): move unit spec to file-ops
- skills/_scripts/libs/file-io/__tests__/unit/backup-old-path.unit.spec.ts:
  file-io 配下の旧 unit spec を削除（file-ops へ移動）
- skills/_scripts/libs/file-ops/__tests__/unit/backup-old-path.unit.spec.ts:
  file-ops 配下に unit spec を新規追加。既存の "too many backups" エラーケース (T-LIB-B-05-01) を
  `When: 異常系` ブロックに再構成し、`StatProvider` が `Deno.errors.NotFound` を reject する
  ファイル不在時の正常終了ケース (T-LIB-B-06-01) を追加。`_fakeStatNotFound` ヘルパーを定義
